+++ types/decimal.go
@@ -34,7 +34,7 @@ const (
 	maxDecBitLen = maxBitLen + decimalTruncateBits
 
 	// max number of iterations in ApproxRoot function
-	maxApproxRootIterations = 100
+	maxApproxRootIterations = 300
 )
 
 var (